requirements.txt
bcrypt==3.1.7
beautifulsoup4==4.8.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
Django==1.9.5
django-bootstrap4==1.1.0
django-cors-headers==3.1.1
django-crispy-forms==1.6.0
django-extensions==2.2.1
django-filter==0.13.0
django-markdown==0.8.4
django-pagedown==0.1.1
django-simple-history==1.9.1
djangorestframework==3.3.3
httplib2==0.17.0
idna==2.9
Jinja2==2.11.1
Markdown==2.6.6
markdown2==2.3.1
MarkupSafe==1.1.1
paramiko==2.7.1
passlib==1.7.2
Pillow==6.1.0
psycopg2==2.8.3
pycparser==2.20
PyNaCl==1.3.0
pytz==2019.2
PyYAML==5.3.1
requests==2.23.0
six==1.12.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.8
